+++ native-windows-wysiwyg/src/gui/controls.rs
@@ -1,2 +1,5 @@
 mod labeled_field;
 pub use labeled_field::LabeledField;
+
+mod left_button;
+pub use left_button::LeftButton;