+++ changelog.md
@@ -2,7 +2,7 @@
 * Fixed `Frame` control `set_size`
 * Added `background_color` to `Frame`
 * Added `track_mouse_leaving`, `track_mouse_hover`, `track_mouse_cancel`, and `track_mouse_query` to `GlobalCursor`
-
+* Added `CustomEvent` control to define custom events in a UI
 
 1.0.10
 * Fixed a compiling bug when using `no-default-features`